+++ src/test/java/com/github/_1c_syntax/mdclasses/mdo/metadata/MetadataStorageTest.java
@@ -35,15 +35,15 @@ class MetadataStorageTest {
     var meta = MetadataStorage.getStorage();
     assertThat(meta)
       .isNotEmpty()
-      .hasSize(51);
+      .hasSize(54);
 
     // проверка на ошибки и дубли
     assertThat(meta.values().stream().map(Metadata::type).distinct()).hasSize(meta.size());
     assertThat(meta.values().stream().map(Metadata::name).distinct()).hasSize(meta.size());
     assertThat(meta.values().stream().map(Metadata::nameRu).distinct()).hasSize(meta.size());
-    assertThat(meta.values().stream().map(Metadata::groupName).distinct()).hasSize(48);
-    assertThat(meta.values().stream().map(Metadata::groupNameRu).distinct()).hasSize(48);
-    assertThat(meta.values().stream().map(Metadata::xmlScope).filter(XMLScope.ALL::equals)).hasSize(50);
+    assertThat(meta.values().stream().map(Metadata::groupName).distinct()).hasSize(50);
+    assertThat(meta.values().stream().map(Metadata::groupNameRu).distinct()).hasSize(50);
+    assertThat(meta.values().stream().map(Metadata::xmlScope).filter(XMLScope.ALL::equals)).hasSize(53);
     assertThat(meta.values().stream().map(Metadata::xmlScope).filter(XMLScope.DESIGNER::equals)).hasSize(1);
     assertThat(meta.values().stream().map(Metadata::xmlScope).filter(XMLScope.EDT::equals)).isEmpty();
   }
